GH-8 refactored tests into their own modules following app structure
