+++ tests/common/mod.rs
@@ -1,1 +1,2 @@
-pub mod utils;
+/// test utils folder
+pub mod test_utils;
