--- tests/healthz.rs
@@ -1,20 +1,0 @@
-//! tests/healthz.rs
-pub mod common;
-use common::utils::spawn_app;
-
-#[tokio::test]
-async fn healthz_works() {
-    // Arrange
-    let address = spawn_app();
-    let client = reqwest::Client::new();
-
-    // Act
-    let response = client
-        .get(format!("{}/healthz", &address))
-        .send()
-        .await
-        .expect("Failed to execute request.");
-    // Assert
-    assert!(response.status().is_success());
-    assert_eq!(Some(0), response.content_length());
-}
