+++ tests/models/mod.rs
@@ -1,0 +1,2 @@
+/// All tests related to the models used in the app
+pub mod shapes_tests;
