+++ tests/routes/api_tests.rs
@@ -1,6 +1,21 @@
-//! tests/subscriptions.rs
-pub mod common;
-use common::utils::spawn_app;
+use crate::common::test_utils::spawn_app;
+
+#[tokio::test]
+async fn healthz_works() {
+    // Arrange
+    let address = spawn_app();
+    let client = reqwest::Client::new();
+
+    // Act
+    let response = client
+        .get(format!("{}/healthz", &address))
+        .send()
+        .await
+        .expect("Failed to execute request.");
+    // Assert
+    assert!(response.status().is_success());
+    assert_eq!(Some(0), response.content_length());
+}
 
 #[tokio::test]
 async fn subscribe_returns_a_200_for_valid_form_data() {
