+++ tests/tests_main.rs
@@ -1,0 +1,6 @@
+pub mod app;
+pub mod common;
+pub mod models;
+/// This is the main test file for the application
+/// Declare all test submodules here
+pub mod routes;
